fixed circular maven dependencies
